\name{SunsetBG}
\docType{data}
\alias{SunsetBG}

\title{Snake background image (sunset).}

\description{
Contains pixel data for a background image.
}

\usage{
SnakeBG
}

\format{
	A matrix of 200 x 200 pixels, containing data as returned by \code{link{rgb}}.
}

\source{
Various web sources (lorem pixum).
}

\author{
Carsten Croonenbroeck
}

\keyword{SunsetBG}
